refactor(timetable): extract event form state container into use-event-form hook
Behavior-preserving extraction of the entire state/logic layer from
timetable-event-modal.tsx (3008 -> 1269 lines) into
event-form/use-event-form.ts (1964 lines):

- form state, reset effect, reference-data fetches (rooms, categories,
  groups, students, staff incl. retry/stale protection)
- both debounced probes (conflict-check 500ms; shift-coverage 500ms
  with AbortController/5s timeout, canCheckShiftCoverage gate) and
  checkCoverageBeforeSave
- validateForm, body builders, re-plan/materialize orchestration
  (56-day chunks)
- handleSubmit, the scope flow (handleScopeSelect/performSeriesEdit/
  confirmLostEdits/series delete) and the full #1875 lost-edits chain

The modal consumes the hook; its JSX render tail is byte-identical to
HEAD (verified by line-level comparison). The extracted logic region
differs from the original in exactly three lines: useModal() stays in
the modal (UI-layer concern), React.FormEvent becomes an imported
FormEvent type (no UMD namespace in .ts), and QuickRepeatPreset stays
unexported (no consumer; knip).

timetable-event-modal.test.tsx and timetable-api.test.ts are untouched;
all 106 tests pass unchanged, pnpm run check is clean.
